Refactor: Extract widget components and improve theming
- Moved `QuickActionButtons`, `WidgetTheme`, and `ElapsedTimeChronometer` into a new `widget.components` package.
- Extracted the `PowerButton` into its own composable file.
- Introduced `WLEDWidgetTheme` to encapsulate device-state-based theming logic.
- Updated the theme generation to use different palette styles based on the device's online and power state (`Fidelity`, `TonalSpot`, `Neutral`).
- Changed the color of the elapsed time chronometer to use `onSurfaceVariant` for better visibility.
